+++ CHANGELOG.rst
@@ -7,6 +7,15 @@ Changelog
 .. NOTE: This document is user facing. Please word the changes in such a way
 .. that users understand how the changes affect the new version.
 
+0.2.0
+--------------------
++ Add ability for filtering paired sequences.
++ The minimum dnaio version requirement was bumped to 0.9.0.
++ Drop Python 3.6 support as Python 3.6 is end of life.
++ Filters are now applied by command flags. This is easier to understand than
+  the previous method (which selected methods using strings). It is also
+  easier to program and document.
+
 0.1.1
 --------------------
 + Various documentation fixes